fix last issues

+++ WatchCord WatchKit Extension/Utilities/Constants.swift
@@ -94,6 +94,6 @@ final class AccordCoreVars {
 #warning("TO DO")
 func generateFakeNonce() -> String {
     let date: Double = Date().timeIntervalSince1970
-    let nonceNumber = (Int(date)*1000 - 1420070400000) * 4194304
+    let nonceNumber = (Int64(date)*1000 - 1420070400000) * 4194304
     return String(nonceNumber)
 }
